refactored Settings page
