+++ ReleaseNotes.md
@@ -6,3 +6,5 @@
 ### Bug Fixes + Miscellaneous
 - Made the header tab buttons sticky + added an effect to make it stand out when scrolling down
 - Some misc. css + ts adjustments as usual
+- Massively refactored home page
+- Massively refactored settings page
